Move the 'ingest/update/process' starcheck bits to process.py
