+++ mica/starcheck/__init__.py
@@ -1,1 +1,1 @@
-from .starcheck import get_starcheck_catalog, get_starcheck_catalog_at_date, main, get_mp_dir, get_monitor_windows
+from .starcheck import get_starcheck_catalog, get_starcheck_catalog_at_date, get_mp_dir, get_monitor_windows
