+++ scripts/update_starchecks.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
 
-import mica.starcheck
-mica.starcheck.main()
+import mica.starcheck.process
+mica.starcheck.process.main()
